requirements.txt
python-telegram-bot==13.7
requests==2.26.0
beautifulsoup4==4.10.0
aiohttp==3.7.4
